test init
