+++ backend/server.js
@@ -3,18 +3,7 @@ require("dotenv").config();
 const express = require("express");
 const mongoose = require("mongoose");
 
-const workoutRoutes = require("./routes/workouts");
-
-const app = express();
-
-app.use(express.json());
-
-app.use((req, _res, next) => {
-  console.log(req.path, req.method);
-  next();
-});
-
-app.use("/api/workouts", workoutRoutes);
+const app = require("./app");
 
 mongoose
   .connect(process.env.MONGO_URI)
